refactor: move mocket.mode to mocket.core.mode
