+++ mocket/core/mocketizer.py
@@ -1,5 +1,5 @@
 from mocket.core.mocket import Mocket
-from mocket.mode import MocketMode
+from mocket.core.mode import MocketMode
 from mocket.utils import get_mocketize
